packages/core/pyproject.toml
[project]
name = "liftwork-core"
version = "0.1.0"
description = "Liftwork shared core: db, telemetry, models, settings, security"
requires-python = ">=3.12"
license = { text = "Apache-2.0" }
dependencies = [
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "structlog>=24.4",
    "bcrypt>=4.2",
    "PyJWT>=2.10",
    "opentelemetry-api>=1.28",
    "opentelemetry-sdk>=1.28",
    "opentelemetry-exporter-otlp>=1.28",
    "opentelemetry-instrumentation-sqlalchemy>=0.49b0",
    "opentelemetry-instrumentation-asyncpg>=0.49b0",
    "prometheus-client>=0.21",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/liftwork_core"]
